format+cleanup

+++ umlet-elements/src/main/java/com/baselet/element/GridElementUtils.java
@@ -27,10 +27,10 @@ public class GridElementUtils {
 		int x2 = Integer.MIN_VALUE;
 		int y2 = Integer.MIN_VALUE;
 		for (GridElement ge : gridElements) {
-			x = Math.min( (int)((double) ge.getRectangle().getX()*scaling), x);
-			y = Math.min((int)((double) ge.getRectangle().getY()*scaling), y);
-			x2 = Math.max((int)((double) ge.getRectangle().getX2()*scaling), x2);
-			y2 = Math.max((int)((double) ge.getRectangle().getY2()*scaling), y2);
+			x = Math.min((int) (ge.getRectangle().getX() * scaling), x);
+			y = Math.min((int) (ge.getRectangle().getY() * scaling), y);
+			x2 = Math.max((int) (ge.getRectangle().getX2() * scaling), x2);
+			y2 = Math.max((int) (ge.getRectangle().getY2() * scaling), y2);
 		}
 		return new Rectangle(x, y, x2 - x, y2 - y);
 	}
